LayoutTests/media/modern-media-controls/slider/slider-fill.html

<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/slider.css" type="text/css" media="screen">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/slider.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>Slider</code> fill property.");

window.jsTestIsAsync = true;

const slider = new Slider;
slider.width = 200;
slider.value = 0.5;

scheduler.frameDidFire = function()
{
    shouldBeEqualToString("slider.children[0].element.style.width", "100px");
    finishJSTest();
}

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
