LayoutTests/media/modern-media-controls/slider/slider-fill-expected.txt

Testing the Slider fill property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS slider.children[0].element.style.width is "100px"
PASS successfullyParsed is true

TEST COMPLETE
